LayoutTests/http/tests/xmlhttprequest/access-control-preflight-async-header-denied-expected.txt

CONSOLE MESSAGE: Request header field X-NON-STANDARD is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?test=access-control-preflight-async-header-denied&state=header due to access control checks.
PASS: Request successfully blocked.
